Refactored importer integration tests into multiple suites.
Fixes #140.
